Fix minor issue. Move declaration of CriterionType.
