+++ .github/workflows/unit-test.yml
@@ -59,8 +59,6 @@ jobs:
           echo "Running checkpointing tests..." &&
           bash ./tests/checkpointing/test_checkpointing.sh &&
           echo "Running distributed training tests..." &&
-          cd tests &&
           PYTHONPATH=.. python run_dist_tests.py &&
-          cd .. &&
           echo "While running unit tests, the following datasets were downloaded:" &&
           ls ~/.avalanche/data
