+++ avalanche/training/supervised/er_aml.py
@@ -13,7 +13,7 @@ from avalanche.training.plugins.evaluation import (
 )
 from avalanche.training.storage_policy import ClassBalancedBuffer
 from avalanche.training.templates import SupervisedTemplate
-from avalanche.training.templates.base_sgd import CriterionType
+from avalanche.training.templates.strategy_mixin_protocol import CriterionType
 from avalanche.training.utils import cycle
